# docs/openapi/admin.yaml
openapi: 3.0.3
...
info:
  title: voxeltoad — Admin API
  version: 0.1.0
  description: >
    Management-plane REST API for providers, models, routes, plugins, tenants,
    API keys, quotas, usage, and audit. Consumed by the Control Panel UI and the
    generated TypeScript admin SDK.
servers:
  - url: /
    description: Same-origin (admin plane root)
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/quotas:
    get:
      tags: [quotas]
      summary: Read a quota balance by scope
      parameters:
        - name: scope
          in: query
          required: true
          schema: { type: string }
          description: >
            Quota scope key, e.g. "tenant:acme", "group:acme/team-a",
            "key:sk-abc". tenant-admin may only query scopes within its tenant.
      responses:
        '200':
          description: Quota balance
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuotaBalance' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/quotas/topup:
    post:
      tags: [quotas]
      summary: Atomically increment a quota balance (super-admin)
      description: >
        Adds delta micro-units to the scope's balance in a single SQL statement
        (balance += delta), never clobbering concurrent hot-path debits.
        Audited as resource_type=quota.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TopUpRequest' }
      responses:
        '200':
          description: New balance after top-up
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuotaBalance' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ---------------------------------------------------------------------------
  # Overview dashboard. super-admin only.
  # ---------------------------------------------------------------------------
...
  /api/v1/operators:
    get:
      tags: [auth]
      summary: List operators (super-admin)
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Operator list (non-secret; never includes the password hash)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperatorList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [auth]
      summary: Create an operator (super-admin)
      description: >
        Creates a super-admin (no tenant_id) or a tenant-admin (tenant_id
        required). The password is argon2id-hashed server-side; the response
        never contains the password or hash. Audited as resource_type=operator.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOperatorRequest' }
      responses:
        '201':
          description: Operator created (non-secret view)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operator' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/operators/{id}:
    delete:
      tags: [auth]
      summary: Delete an operator and revoke its sessions (super-admin)
      description: >
        Deletes the operator and revokes all its sessions immediately. Refuses
        to delete the last super-admin (409), to keep the platform administrable.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Refused — cannot delete the last super-admin
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    put:
      tags: [auth]
      summary: Update an operator (super-admin)
      description: >
        Update email, password, and/or tenant_id for an operator. At least one
        field must be provided. Empty strings are ignored (not updated). Email
        must be unique; duplicate email returns 400/409.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateOperatorRequest' }
      responses:
        '200':
          description: Updated operator
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operator' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/operators/me/password:
    post:
      tags: [auth]
      summary: Change the authenticated operator's own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string }
      responses:
        '204': { description: Password changed }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ---------------------------------------------------------------------------
  # Tenant-scoped groups (tenant-admin). CRUD + enable/disable toggle (mirrors
  # tenant pattern). Delete is rejected when api_keys reference the group (409).
  # Writes are audited as resource_type=group (ADR-0017 §5).
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque operator session token from POST /auth/login.
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
      description: Opaque keyset cursor from a prior response's next_cursor. Omit for the first page.
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
      description: Max items to return.
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing/invalid/expired session
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not authorized for this resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found (or not in caller's tenant)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type]
          properties:
            message: { type: string }
            type:
              type: string
              description: OpenAI-compatible error type.
              enum:
                - invalid_request_error
                - authentication_error
                - permission_error
                - not_found
                - rate_limit_error
                - api_error
    ListEnvelope:
      type: object
      description: Uniform list envelope. next_cursor="" means no further pages.
      required: [data, next_cursor]
      properties:
        data:
          type: array
          items: {}
        next_cursor:
          type: string
...
    Operator:
      type: object
      description: Non-secret operator view — never includes the password hash.
      required: [id, email, role]
      properties:
        id: { type: integer, format: int64 }
        email: { type: string, format: email }
        role:
          type: string
          enum: [super-admin, tenant-admin]
          deprecated: true
          description: Legacy role text. Prefer role_id for Phase-2 RBAC.
        role_id:
          type: integer
          format: int64
          description: FK to roles table (Phase-2 RBAC).
        scope_kind:
          type: string
          enum: [global, tenant]
          description: Whether the operator's role is global or tenant-scoped.
        permissions:
          type: array
          items: { type: string }
          description: Loaded permission set for this operator's role.
        tenant_id:
          type: integer
          format: int64
          nullable: true
          description: Bound tenant; null for a global-scope operator.
        tenant_name:
          type: string
          nullable: true
          description: >
            Resolved tenant name (for quota scope construction). null for
            super-admin.
    CreateOperatorRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
        role:
          type: string
          enum: [super-admin, tenant-admin]
          description: Legacy role text. Use role_id for custom roles.
        role_id:
          type: integer
          format: int64
          description: FK to roles table (Phase-2). Takes precedence over role.
        tenant_id:
          type: integer
          format: int64
          description: Required for tenant-scoped roles; must be omitted for global roles.
    UpdateOperatorRequest:
      type: object
      description: >
        At least one of email, password, role_id, or tenant_id must be provided.
        Empty strings mean "leave unchanged"; nil means "leave unchanged".
      properties:
        email: { type: string, format: email }
        password: { type: string }
        role_id:
          type: integer
          format: int64
          description: Change operator's role (Phase-2 RBAC).
        tenant_id:
          type: integer
          format: int64
          nullable: true

    # --- API keys ---
...
    QuotaBalance:
      type: object
      required: [scope, balance, currency]
      properties:
        scope: { type: string }
        balance: { type: integer, format: int64, description: Micro-units. }
        currency: { type: string }
    TopUpRequest:
      type: object
      required: [scope, delta, currency]
      properties:
        scope: { type: string }
        delta: { type: integer, format: int64, description: Micro-units to add (must be > 0). }
        currency: { type: string }

    # --- Overview ---
...
    OperatorList:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            data: { type: array, items: { $ref: '#/components/schemas/Operator' } }
...